libmrc/xdmf_collective: a bit more

+++ src/libmrc/src/mrc_io_xdmf_collective.c
@@ -933,7 +933,7 @@ collective_send_fld_begin(struct collective_m3_ctx *ctx, struct mrc_io *io,
       if (!has_intersection)
 	continue;
 
-      size_t len = m3->_dims.vals[0] * m3->_dims.vals[1] * m3->_dims.vals[2];
+      size_t len = (size_t) m3->_dims.vals[0] * m3->_dims.vals[1] * m3->_dims.vals[2];
       buf_sizes[writer] += len;
     }
 
@@ -1143,7 +1143,7 @@ collective_recv_fld_end(struct collective_m3_ctx *ctx,
       continue;
     }
     
-    size_t len = (size_t)(ihi[0] - ilo[0]) * (ihi[1] - ilo[1]) * (ihi[2] - ilo[2]);
+    size_t len = (size_t) (ihi[0] - ilo[0]) * (ihi[1] - ilo[1]) * (ihi[2] - ilo[2]);
 
     //mprintf("B recv_fld_end gp %d\n", gp);
     switch (mrc_fld_data_type(m3)) {
